refactor: priority 系関数を移動
