+++ voicevox_engine/user_dict/part_of_speech_data.py
@@ -2,6 +2,7 @@ from enum import Enum
 from re import findall, fullmatch
 from typing import Any
 
+import numpy as np
 from pydantic import BaseModel, Field, validator
 
 USER_DICT_MIN_PRIORITY = 0
@@ -267,3 +268,25 @@ class UserDictInputError(Exception):
     """受け入れ不可能な入力値に起因するエラー"""
 
     pass
+
+
+def _search_cost_candidates(context_id: int) -> list[int]:
+    for value in part_of_speech_data.values():
+        if value.context_id == context_id:
+            return value.cost_candidates
+    raise UserDictInputError("品詞IDが不正です")
+
+
+def _cost2priority(context_id: int, cost: int) -> int:
+    assert -32768 <= cost <= 32767
+    cost_candidates = _search_cost_candidates(context_id)
+    # cost_candidatesの中にある値で最も近い値を元にpriorityを返す
+    # 参考: https://qiita.com/Krypf/items/2eada91c37161d17621d
+    # この関数とpriority2cost関数によって、辞書ファイルのcostを操作しても最も近いpriorityのcostに上書きされる
+    return MAX_PRIORITY - np.argmin(np.abs(np.array(cost_candidates) - cost)).item()
+
+
+def _priority2cost(context_id: int, priority: int) -> int:
+    assert MIN_PRIORITY <= priority <= MAX_PRIORITY
+    cost_candidates = _search_cost_candidates(context_id)
+    return cost_candidates[MAX_PRIORITY - priority]
